#236: EditToDoSceneTheme 추가
- UI와 관련된 데이터들을 저장하는 Theme 객체를 추가하였습니다.
- 기존의 레이아웃 관련 상수들을 저장한 Constant 선언부를 Theme 내부로 이동하였습니다.
